Rename test_data to testdata for consistency.
We sometimes treat the "testdata" directory specially, e.g.,
excluding it from the .par, so we need to be consistent about
its naming.  The umpire tests sometimes modify files inside
'testdata', so we need to get this right or the rsync operation
in make_par may fail when tests are run in parallel.

This also fixes a big in run_tests.py that caused "make test" to
fail because files had been renamed.

BUG=None
TEST=Unit tests

Change-Id: If918ea9d1ad9e67a341040dfd8c3e250da78ea54
Reviewed-on: https://chromium-review.googlesource.com/210892
Tested-by: Jon Salz <jsalz@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Commit-Queue: Jon Salz <jsalz@chromium.org>