touchscreen_calibration: get initial sensor status
Due to the fact that the method of checking the sensor status has
been changed, in order to correctly get the sensor status when
Arduino is booted, we need to update the sensor status, wait a little
while, and then update the sensor status again.

Without this patch, if the probe parks at the UP position initially,
it may mistakenly think it is not at the UP position. The result is
that the probe will move up until it reaches the EXTREME UP position.
This behavior may be dangerous. If Arduino is reset one more time,
the probe will try to move from EXTREME UP to even higher and hit
the ceiling of the test fixture.

BUG=chrome-os-partner:31027
TEST=Manually test it on a touchscreen calibration test fixture.
Initially, park the probe at the UP position by pressing the debug
button. Use a host to connect to the Arduino. Assume that the
programming port is /dev/ttyACM0. Reset the Arduino a couple of
times by using screen utility:

$ screen -L /dev/ttyACM0 9600

Observe that the probe will keep parking at the UP position instead
of moving up.

Change-Id: Ide69dda6854662a18aeb6867866f336cb67ba8b5
Reviewed-on: https://chromium-review.googlesource.com/211066
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>