Tune down link_manager logging
If the controller sees the presenter, but the presenter doesn't see the
controller, the controller would keep trying to get to the presenter and
fail. This produces a large amount of log text. Let's tune it down so
that for each presenter/controller IP, we only report it once in the log.

BUG=chrome-os-partner:30202
TEST=Run in the environment stated in the commit message. Check the log.

Change-Id: Id94b14c9e95083203a3a3924255df33c27e0eb6c
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/211038