touchscreen_calibration: refactoring fixture code as an individual module
As we will have a new generation of test fixture, the code about
fixture manipulation is moved out as an individual module.

Current test fixture is a semi-automated one, and the next one will
be more manual without auto moving parts. It would be cleaner to
separate the fixture from the test code.

BUG=chrome-os-partner:31264
TEST=manually verified the refactoring code on the old test fixture.

Change-Id: I3072ee2275b7d5b259972210ce43e29f6df2e9d9
Reviewed-on: https://chromium-review.googlesource.com/211781
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>