touchscreen_calibration: refactoring using XML RPC client-server architecture
Previously, when testing an AB panel, it requires the MLB or the
assembled unit of the same platform of the panel. However, such
MLB or assembled unit may not be always available due to factory
inventory and management issue.

A solution is to use a Beagle Bone to probe the touchscreen sensors,
and use a standard host to display the factory GUI.

In this patch, the sys fs related functions are separated into an
individual module which could run as an XML RPC server. The test
itself acts as an XML RPC client which connects to the server for
sys fs related services.

This client-server architecture is very flexible in that the client
and the server could run on the same or different machines depending
on what scenario is adopted. In particular, the server could be run
on the same machine as the client, the Beagle Bone, or even on a DUT.

BUG=chrome-os-partner:31264
TEST=manually verified the refactoring code with the old test fixture
for two scenarios:
(1) Both the client and the server run on the same host.
(2) The client runs on a host, and the server runs on a Beagle Bone.

Change-Id: Icbb785e44b53c6fff2817e336f5dc6a1fd775cdb
Reviewed-on: https://chromium-review.googlesource.com/216266
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>