Accelerometer: Fix reading raw data fail
We expect to read 12 bytes raw data from the sysfs but sometimes
we couldn't read it. It may be related to Chrome is getting the data at the
same time. The correct solution is to tell Chrome stop reading the raw
data during calibration. However, we don't know how to do it right now.
So using retry as a short term solution.

BUG=chrome-os-partner:31650
TEST=manually test on Glimmer

Change-Id: Ib735d2500b65e8d352cf796af0fa6f4667587aad
Reviewed-on: https://chromium-review.googlesource.com/214454
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>