touchscreen_calibration: set iptables for dport 8000
This patch enables the destination port of tcp input traffic on
the machine running the sysfs_server so that a client can connect
to the server.

BUG=chrome-os-partner:31264
TEST=manually verified the code with the old test fixture.
To verify the iptables has been properly setup, let the client
runs on a host, and the server runs on a Beagle Bone. Observe
that the client can connect to the server without problem.

Also verified that both the client and the server can run on the
same host.

Change-Id: I57cc36d32c0d4c2c56b1d80908cc774288dd4ced
Reviewed-on: https://chromium-review.googlesource.com/216267
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>