touchscreen_calibration: get sys fs paths
If the sys fs and debug fs are not specified in the conf file,
probe their paths automatically. This is handy since the paths
may keep changing during different build phases.

The entries in the conf, if set, have the higher priority.

BUG=chrome-os-partner:31264
TEST=manually verified the code with the old test fixture
for two scenarios:
(1) Both the client and the server run on the same host.
(2) The client runs on a host, and the server runs on a Beagle Bone.

Change-Id: I2b1c80d90109a12a9857bc02a233cdb9487185a4
Reviewed-on: https://chromium-review.googlesource.com/216268
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>