touchscreen_calibration: get sysfs service with different scenarios
It is possible to run the client and the server on the same or
different machines. On the client side, it should be able to
determine where to get the sysfs service. This could be done
by configuring the generic_tsab test list properly.

If the sysfs_server runs as a process on a different machine,
the _SYSFS_SERVER_IP should be assigned the IP address of the server
machine. The client would then try to connect to the server as
a first priority.

If the _SYSFS_SERVER_IP is not assigned, the client will then assume
that the client machine is directly connected to the DUT. Hence, the
client will try to instantiate a Sysfs object for the sysfs service.

BUG=chrome-os-partner:31264
TEST=manually verified the refactoring code with the old test fixture
for two scenarios:
(1) Both the client and the server run on the same host.
(2) The client runs on a host, and the server runs on a Beagle Bone.

Change-Id: I56045431b0b1ed1df8612c3efde1545de42f1832
Reviewed-on: https://chromium-review.googlesource.com/216269
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>