touchscreen_calibration: add use_sysfs_server flag in the configuration
A sysfs service is used to provide sensor data.

This patch allows a user to configure either to use the sysfs_server
on a remote machine or on the same host as a sysfs object. This is
a more deterministic way to launch the service.

In either way, if a sysfs service cannot be acquired, display the
fatal error in a pop-up message.

BUG=chrome-os-partner:33078
TEST=manually verified with a test host and a Beagle Bone setup
on a test fixture.

Change-Id: Iab0ebdf5c93919f84cd797feaca5fef758e20189
Reviewed-on: https://chromium-review.googlesource.com/223904
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>