Make goofy and goofy_presenter less likely to hang after SIGINT
When the user presses control-C, goofy raises a KeyboardInterrupt
exception. This is caught and begins the shutdown process, which
can take several seconds. If the user is impatient and presses
control-C again, the shutdown process itself gets interrupted,
and goofy.py completes. However, there are likely to be threads that
are still running, and Python's Thread class installs an atexit
handler that waits for all (non-daemon) threads to exit. Since
the shutdown process was interrupted, the threads will never exit,
and so the Python Thread class will not allow the goofy process to
exit. goofy appears to hang.

This fix makes goofy_base take responsibility for thread teardown.
This allows us to log the threads that we are waiting for (useful
for debugging), and also support a forcible exit. If the user
presses control-C while we are waiting for threads to exit, goofy
now exits abruptly, without invoking Python's atexit handler. This
means that one control-C invokes graceful shutdown, while a second
control-C will always exit, even if goofy is wedged.

BUG=none
TEST=make test, manual

Change-Id: I94b0cd930e422b3f04d36aab61901bc0db2af4b7
Reviewed-on: https://chromium-review.googlesource.com/219514
Commit-Queue: Peter Ammon <pca@google.com>
Tested-by: Peter Ammon <pca@google.com>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>