Preserve the test UI when presenter reconnects
When the connection between the presenter and the device drops, just
hide the UI instead of destroying it. Upon reconnection, compares the
UUID which is randomly generated for each UI session, and reuses the UI
from the last connection if it's the same session. This way, we can sort
of preserve the UI of the active test. However, if some events are lost
when the connection is broken, the UI might be in an inconsistent state.
Still, this change should help us in most cases.

BUG=chrome-os-partner:32670
TEST=Run a test with UI. Unplug the Ethernet cable and see the UI goes
away. Re-plug the cable and see the UI of the active test.

Change-Id: I37a0d3c95b32f155e043ea2710c1b1b866457b4c
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222942
Reviewed-by: Hung-Te Lin <hungte@chromium.org>