serial_utils : catch SerialException for particular usage
During sending of pyserial, SerialException will raise if the connection
is suddenly broke in the middle of command sending, ex. the sending
command will cause the hardware mux on the connection switch and cut out
the connection.

We separate this to SerialTimeoutException since we need to handle them
as two different cases.

BUG=chrome-os-partner:29166
TEST=update unittest. Test sending normal command and cut-out command
on Samus with fixture.

Change-Id: I316c51057cf683e0fa7687fa052b5e1d590cdc21
Reviewed-on: https://chromium-review.googlesource.com/223230
Reviewed-by: Dean Liao <deanliao@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>