touchscreen_calibration: use serial number 0000 as experimental tests
In order to prevent factory operators from scanning wrong bar codes
in the factory, there is a length check on the serial number.
Currently, the length is 22 which is a pretty long one. For the
development purpose, if a special number, i.e., 0000, is input, it
indicates that an experimental test is conducted, and the results
should be ignored.

BUG=chrome-os-partner:33107
TEST=manually verified in the factory.

Change-Id: I0b3e2dc0d514dd6984a9d0e06803371608eb03b0
Reviewed-on: https://chromium-review.googlesource.com/232171
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>