factory test: add module to pull in autotest libraries
There are some libraries in the autotest tree that are useful for
factory tests.  To make use of them, we need a helper module to pull
in dependencies from the correct autotest directory.  Its location
can vary depending on whether the test is running from a DUT on a
test image or from a development chroot.

BUG=None
TEST=Manual test

Change-Id: I25d12249994b2df26237cdcdcfaaaf03ff15a550
Reviewed-on: https://chromium-review.googlesource.com/224241
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Joel Kitching <kitching@google.com>
Commit-Queue: Joel Kitching <kitching@google.com>