Respond to presenter/controller pings in a separate process
The presenter and the controller ping each other periodically to
determine if the link is still good. The pings are currently processed
in the same threads the link managers are in. On slow devices, if Goofy
is busy, the pings get processed slowlier and sometimes it's too slow
that the presenter thinks the controller is down.

Avoid this by starting a standalone "ping server", which is only
responsible for responding to pings.

BUG=chrome-os-partner:33226
TEST=Run controller on Beaglebone. Doesn't see broken connection as
often.

Change-Id: I2984c396595641e7d5f5cc1bbd42ea0b6a57787e
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226640