Legacy registration code: raise error if user/group code are the same
Currently we check VPD user/group bind values by their format, but not
checking their actual data. It was found that some devices were assigned
with same user/group values and that should be apparently wrong value,
according to their definition. To catch that failure, we want to add a
simple test to assert when user/group codes are the same.

This is only required for legacy format as we will check the type (user
or group) for new-style codes.

BUG=chrome-os-partner:33473
TEST=Manual
TEST=unit test

Change-Id: I672335012a5a27eb579927181df5feeea278a94c
Reviewed-on: https://chromium-review.googlesource.com/227271
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Stephen Lin <stephenlin@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Bowgo Tsai <bowgotsai@google.com>