Check all files in unencrypted are own by root:root.
This patch verifies that all the UID:GID for the files included in
the stateful_files.tar.xz are own by root. This prevents to introduce
a bug due to having an unencrypted directory with wrong metadata when
creating the stateful_files.tar.xz in the factory.

BUG=chromium:432774
TEST=Run the new lines in the chroot (just the new part) and tested that locally.

Change-Id: Iebfaa914bd6944e597e7e6596e6d3cfdf9ca2f3d
Reviewed-on: https://chromium-review.googlesource.com/229967
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>