touchscreen_calibration: add use_shopfloor flag
In the development environment, the shopfloor server is not always up
and running. In that case, the script would block waiting endlessly
for the shopfloor server. An alternative method is to define a flag
indicating if the shopfloor server is used. If the flag is false,
just throw away the log and let the script proceed.

BUG=chrome-os-partner:33107
TEST=manually verified with a test fixture.

Change-Id: I8ccc02d5f88e11369b6f382a9923bdf47d4989ba
Reviewed-on: https://chromium-review.googlesource.com/232172
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>