goofy: Merge goofy_monolithic and goofy_split.
Most files in Goofy_split and Goofy_monolithic are either the same or with minor
difference.

Since Goofy_split development is close to completion now (although still
unstable on certain platforms), we want to merge the two implementations so
people can switch easily and developers don't need to worry about updating code
in two different locations.

To get better stability, we want to run as monolithic way when system is running
in standalone mode, and run as goofy-split way when the system is configured to
run only presenter or device.

BUG=none
TEST=Started standalone and presenter-only modes on DUT.
CQ-DEPEND=CL:238141

Change-Id: Ide813b9475e975930d215111cacbfb3b437576fb
Reviewed-on: https://chromium-review.googlesource.com/237927
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>