tools/deps: Clean import space when seeing import failure.
Some packages (e.g., twisted, zope) may try to register statically and would
cause problem when we try to reload them in dependency checker.  We used to
workaround that by pre-importing but that does not work well when the key module
can't be determined.

A better approach is to restart (by exec) checker with a clean import space, and
retry one time. This should work on all packages using same approach for
registration in future.

BUG=chromium:446394
TEST=py/tools/deps.py py/umpire/*.py # pass

Change-Id: I199551771dc885ec66e0a7e58f1ce7430dfd1322
Reviewed-on: https://chromium-review.googlesource.com/238664
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>