Add callback hooks for test start and test failure
For headless device, we need a way to notify operator of test status
when the devices are on the run-in rack, which can be done with these
hooks. For example, light up green LED when tests start and red LED when
a test fails. The operators can then easily tell which device fails by
simply looking at the LED.

BUG=chrome-os-partner:35023
TEST=Check the callbacks are called:
       1. Start goofy -> OnTestStart called.
       2. Manually swich/restart test -> OnTestStart called.
       3. Test fails -> OnTestFailure called.
       4. Abort a test -> OnTestFailure called.

Change-Id: I190e20e5e66fff1fd0ee52324a472c51d3d9795c
Signed-off-by: Vic Yang <victoryang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/238950
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Vic Yang <victoryang@chromium.org>
Tested-by: Vic Yang <victoryang@chromium.org>