Add hooks in 'summary' test
For headless devices, we need a way to notify the operators that some
tests have finished, e.g. run-in tests. To accomodate devices of
different form factors, let's add hooks in Board classes. At the end of
a test section, we can now have a summary test there acting as a blocker
and also invoking OnSummaryGood/Bad hooks.

For example, at the end of run-in, we can add a summary test. And in the
board class, have OnSummaryGood() light up a green LED and
OnSummaryBad() light up a red one. This would allow the operators to
figure out which devices have finished run-in, and it'd be clear which
ones passed/failed.

BUG=chrome-os-partner:35023
TEST=Check the callbacks are called in both pass/fail cases.

Change-Id: Ie7b391c812a417759909e840be545af898d8834c
Signed-off-by: Vic Yang <victoryang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/240286
Tested-by: Vic Yang <victoryang@chromium.org>
Trybot-Ready: Vic Yang <victoryang@chromium.org>
Reviewed-by: Chun-ta Lin <itspeter@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Vic Yang <victoryang@chromium.org>