Handle presenter announcement while connected
On the DUT side, if we receive presenter announcement while the link was
already established, it means the presenter is somehow back into
disconnected state and is trying to establish connection again. In this
case, instead of ignoring the announcement, we should register again
and go back to disconnected state if registration fails. This way,
either both side establish a new link or both go back to disconnected
state.

BUG=chrome-os-partner:35754
TEST=Unit test.

Change-Id: Iab55d8ad540208e13744ac5aa0a7171501990315
Signed-off-by: Vic Yang <victoryang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/243781
Tested-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Vic Yang <victoryang@chromium.org>