led: refactor and update to support darg i18n strings
Refactor led test by subclassing CheckLEDTask to CheckLEDTaskNormal
and CheckLEDTaskChallenge.  Introduce a new darg 'index_i18n', which,
when provided, allows Chinese translations for LED index names to be
specified in a test_list.  If available, these will be shown in the
Goofy UI.

Finally, turn an LED light off after running its test and before
starting the next one.  This prevents the case where the lights come
on one-by-one and stay on as the targets progress.

BUG=chrome-os-partner:36402
TEST=Manually on DUT

Change-Id: Iaed0eaa693fbbad324cd57abb6aee1946b563697
Reviewed-on: https://chromium-review.googlesource.com/248550
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Joel Kitching <kitching@google.com>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>