sys_utils: Fix unit test due to mke2fs default behavior changes.
The owner of root directly in the file system created by mke2fs should be the
user running that command; however on some Linux systems that may be enforced as
root. To workaround that, use -E owner_root=uid:gid so the owner is explicitly
specified.

BUG=none
TEST=make test

Change-Id: I7e53ce04e4fe9b3b01ec55b06547f7a68bac4ed5
Reviewed-on: https://chromium-review.googlesource.com/256256
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Joel Kitching <kitching@google.com>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>