umpire client: Add quiet mode to suppress error message
Add a quiet mode to suppress error message when server can not be
reached.
Also, change a message from warning to debug when request is made by
user but server version is not decided yet. This is common when server
can not be reached.

BUG=chrome-os-partner:37650
TEST=Not used yet.

Change-Id: Iac243cd57389aca7fdbd1c75731428356627cf12
Reviewed-on: https://chromium-review.googlesource.com/260485
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>