goofy_rpc: fix EventClient threads not being terminated
When running RPC calls in goofy/goofy_rpc.py that require CallExtension, a
dedicated EventClient is created.  If this EventClient (which starts a new
thread) is not eventually closed, and if enough calls to CallExtension are made,
the calls will start failing with "error: can't start new thread", with the
thread count reaching upwards of 330.

Closing the EventClient instance after using it fixes this problem.

BUG=None
TEST=Manually on DUT

Change-Id: I69e8e4792c386ac4faa812624ef6a1479f0d7e8f
Reviewed-on: https://chromium-review.googlesource.com/263684
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>