lte_verify_config: allow use of modem_utils.GetModem
Sometimes we can't be sure of an LTE modem's /dev/ttyACMxx
device path, and we have to rely on the GetModem function
implemented in modem_utils to do the work for us.

To allow for this situation, make the modem_path argument
optional, and if not provided, then use modem_utils.GetModem
to create the Modem object.

BUG=chrome-os-partner:37149, chrome-os-partner:38271
TEST=Manually on DUT

Change-Id: I3fdb6ac769a198ce391b9e70f22d4ff6f203bb7d
Reviewed-on: https://chromium-review.googlesource.com/271736
Reviewed-by: Mao Huang <littlecvr@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>