overlord: allow ghost to specify machine ID to use
In fixture-based testing, fixture host does not have MLB number so the
machine ID is a randomly generated UUID string. This make it hard to
identify fixtures. We add a command line switch to allow us to assign a
specific machine ID, so we can find it easily in the overlord dashboard.

BUG=none
TEST=`./go/src/overlord/test/overlord_e2e_unittest.py`

Change-Id: I2930b5f980d4ac1f279c089ea482106eaa1463db
Reviewed-on: https://chromium-review.googlesource.com/272060
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>