call_shopfloor: support action_args in action handler
In the folio_base project, the data returned by shopfloor with
the QueryBtMac method is a tuple of (result, sn, value). In that
project, the shopfloor does not know the mac_address_key. Hence,
the base_host has to pass the mac_address_key in action_args in
order to save the MAC address in the factory shared data.

BUG=chrome-os-partner:42640
TEST=manually tested.

Change-Id: I48b4c03e0e023f1ecacd422c8db803d5f4aa3ea3
Reviewed-on: https://chromium-review.googlesource.com/286707
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>