goofy: Add new RPC call 'ScheduleRestart'.
The Goofy RPC "RestartTests" will terminate (and log) all running tests so
calling that directly in a test (usually as end of fixture tests) would end with
lots of exception logs.

To perform a graceful restart, we want to restart in end of the test.

BUG=chromium:506788
TEST=manually tested.

Change-Id: I820da3b04f0f0f258b32bfd57f8ed7a24a0faee4
Reviewed-on: https://chromium-review.googlesource.com/286750
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>