call_shopfloor: set allow_none when creating ServerProxy
If an operator failed a test, e.g., scanning the operator ID, its
default value may be simply None. When sending the values to the
shopfloor server in the EndBaseFAT method, the server proxy would
raises an exception when seeing the None value. By setting
allow_none to True when creating the server proxy would solve the
problem.

BUG=chrome-os-partner:42640
TEST=manually tested.

Change-Id: I3fb701256744741b1e0c57f815322b23ea7a240b
Reviewed-on: https://chromium-review.googlesource.com/287628
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>