Add methods to verify refs and deltas values before touched
To verify the refs data, we check the uniformity of the refs values.
A NORMALIZED_DEVIATION_THRESHOLD is defined in the config file such
that the normalized deviation of every sensor from the mean should
be less than this threshold.

To verify the deltas values before the conductive foam touching
the panel, a DELTA_UNTOUCHED_HIGHER_BOUND is defined in the config
file such that the delta value of every sensor should be less than
this threshold.

BUG=chrome-os-partner:43044
TEST=manually tested.

Change-Id: I4b29e0825356c068fb377f65ea6d044747a1ba99
Reviewed-on: https://chromium-review.googlesource.com/287520
Reviewed-by: Premal Parekh <parekhp@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>