display: Check fb object before access it
For some external display, the fb object may be assigned as None
right after the HDMI plugged in. It will trigger exception when
we access it with fb.width and cause the test failed. To avoid it,
check the fb before access it.

BUG=chrome-os-partner:41294
TEST=Run ext_display test item and get pass result.

Change-Id: I11b242ecf0b67dc11abfca296c694cc698b4d569
Reviewed-on: https://chromium-review.googlesource.com/291082
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Kevin Cheng <kevin.cheng@intel.com>
Commit-Queue: Kevin Cheng <kevin.cheng@intel.com>