Add an option to retry the failed test on start.
Considering a test flow: when one of the test failed, the DUT is sent to
repair the component. After the DUT is repaired, it should run the
failed test again to verify the repairing process is successful.

The option retry_failed_on_start is for this situation. The failed tests will
automatically run on the start while the option retry_failed_on_start is
enabled. It is only effective when auto_run_on_start is also enabled.

BUG=chrome-os-partner:43592
TEST=manually test on device
     make lint LINT_WHITELIST="py/{goofy/goofy.py,test/factory.py}"

Change-Id: I9024e51291db771cd1fb3b43dc3d01213d2e70dc
Reviewed-on: https://chromium-review.googlesource.com/293970
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>