py/test: add a bluetooth_utils module
Add a bluetooth_utils module to read bluetooth device services
through UUID. Specifically, it supports a wrapper class of gatttool
to read the battery level and the firmware revision string.

This module is mostly inspired by a script rel_tester.py written
by mylesgw@chromium.org. However, rel_tester.py is not actually
submitted to the code base. It is also better to import a module
and invoke its methods instead of spawning a subprocess to execute
the script directly.

BUG=chrome-os-partner:44403
TEST=manually tested.

Change-Id: I9e4949ac49e933996c0ae841f826ccd9da5d111b
Reviewed-on: https://chromium-review.googlesource.com/294696
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>