bluetooth: ReadBatteryLevelTask and CheckBatteryChargeTask
The original ChargeTestTask consists of reading battery level step 1,
reading battery level step 2, and checking if the battery level has
increased over the two steps.

In this patch, the task is split into 3 tasks. Ideally, we could read
the battery level before charging, start charging the battery, conduct
other tests while the battery is charging, read the battery again, and
stop charging. This involves re-ordering of the test list and will be
accomplished in a following patch.

This patch also removes the scan_mac_only arg which is confusing.

BUG=chrome-os-partner:44403
TEST=manually tested.

Change-Id: I5abaa8037acd9f9c385d630e9d77f1ad181156b5
Reviewed-on: https://chromium-review.googlesource.com/301845
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Myles Watson <mylesgw@chromium.org>