touchscreen_calibration: PHASE_SETUP_ENVIRONMENT
Originally, there was only a test in touchscreen_calibration which
included pinging the network status, setting up of its environment,
e.g., pushing required executables into the device under test, etc.
which took a few seconds. We are in an effort of adding more tests
into the test list. It would save substantial time by separating the
task of setting up the environment and only running it once in the
beginning of the test list instead of executing the task repeated
in every test.

BUG=chrome-os-partner:43044
TEST=manually tested.

Change-Id: Id591ccc7ff093603a50e7a42e8eec79976cf7322
Reviewed-on: https://chromium-review.googlesource.com/298400
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Premal Parekh <parekhp@google.com>