Dolphin: Add double CC USB Type-C cable and more parallel testing support
When 'parallel_test' flag is enabled, test flow for Type-C is:
 - Set PR IOExpander to default.
 - Set PR Type-C mode to default.
 - Set PR Type-C mode to USB 3.0.
 - USB 3.0.
 - USB 2.0, 5V/20V, and DP.
 - CC (and AutoFlipCC if double CC USB Type-C cable is in use).
 - 20V.
Note that tests on the same line run in parallel.

Notice that if we are at 20V charge mode and then CC is flipped, after
configuration communication is completed via the new CC pin, the charge mode
will automatically be set to 20V again. Therefore, to verify 2nd CC pin, we'll
just need to check if 20V charge mode is activated again. If double CC USB
Type-C cable is in use, the flipping will be done by the mux on PR. This can
save operators' effort on manually flipping the cable.

The USB Type-A port paired with the Type-C port is used to control PR (USB 2.0)
and receive DP signal converted to webcam video by Plankton-HDMI (USB 3.0). If
these work fine, it means this Type-A port is good.

Specifically, in dolphin_bft_fixture.py, add a function to return if parallel
testing is enabled, a variable double_cc_cable to indicate if double CC USB
Type-C cable is in use, and a function to flip CC mux without changing current
Type-C mode.

In raiden_cc_flip_check.py, when 'parallel_test' flag is enabled, all we want
is to quickly find out if CC polarity is toggled successfully or not. We are
not interested in the current Type-C mode and and the port status. Moreover,
if double_cc_cable is True, the orientation of the double CC cable being
inserted into DUT and PR, and the actual CC polarity being detected initially
(its random), are not important. What we really care is both CC pins are
verified. The verification order is either CC1->Flip->CC2 or CC2->Flip->CC1,
and in either case, both CC pins are verified.

BUG=chrome-os-partner:39472
TEST=Manual test on Glados.

Change-Id: I4d96d9b35b94fa96d77e5b3a1868c7fc31e74ee2
Reviewed-on: https://chromium-review.googlesource.com/295624
Commit-Ready: JAMES C Chen <james.c.chen@intel.com>
Tested-by: JAMES C Chen <james.c.chen@intel.com>
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>