pytests: add spatial sensor calibration test
Calibrate spatial sensors like accelerometer and gyroscope.  This test
is used for some chipsets like BMI160, the calibration is done by
writing a flag in the sysfs entry then read out the calibbias.
This test only requires the device to be placed on a flat service and
trigger the calibration routine in the driver.

BUG=chrome-os-partner:44770
TEST=manually on device

Change-Id: I0451b15616f178f5000e480c88fdd903dcc781a7
Reviewed-on: https://chromium-review.googlesource.com/296916
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>