merge_factory_usbimg.sh: fix incorrect error message
At the end of this script, we'll check the existence of the generated
output image in the trap function. However, the $output is a local
variable inside main() and will become empty after leaving main().
So it will always print the error message that output file is not
generated in the trap function.

Use double quote to expand the local variable for the trap function
before leaving main().

BUG=chrome-os-partner:36725
TEST=./merge_factory_usbimg.sh -f output.bin usbimg1.bin usbimg2.bin

Change-Id: I638cb84f5909ff365cc111699e20c5f1d66b4258
Reviewed-on: https://chromium-review.googlesource.com/297935
Commit-Ready: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>