dkps: record DRM key hashes to prevent from uploading duplicate keys
Some filters may change the content of the DRM keys, even add random factors.
For example, if we want to use a hardware random number generator to encrypt the
keys, the host thats responsible for writing the keys to devices doesnt have
one but the DRM keys provisioning server has one. We therefore need to generate
the random numbers on the DRM keys provisioning server and encrypt the DRM keys
before saving them.

With the example, even with the same original DRM key, the hardware random
generator would generate different keys if the uploader uploads the same key
twice. Thus, to prevent from adding duplicate DRM keys into the database, we
need to record the hash of the original DRM key before filter it.

BUG=chrome-os-partner:45794
TEST=Unit test

Change-Id: I2aca2eb9422e78bc911a63ceb3b5663e1c7e00ba
Reviewed-on: https://chromium-review.googlesource.com/302291
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>