accelerometers_lid_angle: fix angle calculation for kernel >= 3.18
The standard orientation of the two accelerometers was changed to
use the same orientation as that in Android and w3 spec.
Previously Y axis is aligned with the hinge but it's changed to X
axis now. The following is the detailed info.

w3 spec: http://www.w3.org/TR/orientation-event/#description

For kernel version >= 3.18 in ChromeOS, when lid angle is 180 degrees,
both base/lid accelerometers have the same orientation:
  +X axis is aligned with the hinge and pointing to the right.
  +Y axis is in the same plane as the keyboard pointing towards the
     top of the screen.
  +Z axis is perpendicular to the keyboard, pointing out of the keyboard.

In previous kernel versions, when lid angle is 0 degrees, both base/lid
accelerometers have the same orientation:
  +X axis is in the same plane as the keyboard pointing towards the
     top of the keyboard.
  +Y axis is aligned with the hinge and pointing to the right.
  +Z axis is perpendicular to the keyboard, pointing out of the keyboard.

BUG=chrome-os-partner:44973
TEST=Manual with several angles, 45, 90, 180, 270, etc.

Change-Id: I41c80bcdcac38557859c1d08a68a077cac9b4e71
Reviewed-on: https://chromium-review.googlesource.com/303645
Commit-Ready: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>